airflow/requirements.txt
langchain-core
langchain-classic
langchain-community
langchain-docling
langchain-falkordb
langchain-google-genai
langchain-neo4j
langchain-openai
langchain-postgres
langchain-qdrant
langchain-text-splitters

